# Conflicts:
#	src/main/java/com/nhom7/ecommercebackend/controller/ProductController.java
#	src/main/java/com/nhom7/ecommercebackend/request/ProductDTO.java
#	src/main/java/com/nhom7/ecommercebackend/service/impl/ProductServiceImpl.java

+++ src/main/java/com/nhom7/ecommercebackend/model/CartItem.java
@@ -19,9 +19,9 @@ public class CartItem extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-    private int quantity;
+    private Long quantity;
 
     @ManyToOne
     @JoinColumn(name = "card_id")
